ast27x0: refactor FIT boot loader into fmc_image.c for future Caliptra Manifest support
Refactor the BootROM image loader to prepare for future SoC Caliptra manifest
support.

Move FMC scan and FIT parsing/loading logic out of image.c into a dedicated
implementation file (fmc_image.c) and add a corresponding header (fmc_image.h).
Keep image.c focused on boot entry and boot policy flow, including a placeholder
policy step for Caliptra Manifest boot with fallback to the legacy FIT path.

Update headers to include stdint.h where required, and adjust Makefile objects
to build the new module.

This change is a pure refactor with no intended functional behavior change. It
cleans up the code base and provides a clear separation between boot policy and
image-format-specific logic, which will be extended by a future Caliptra
manifest loader.
